Int test refactor: package rename and class moves.
