+++ jvm/workbookapp/src/integration-test/kotlin/integrationtest/projects/RowCount.kt
@@ -1,4 +1,4 @@
-package integrationtest.rcimport
+package integrationtest.projects
 
 import org.wycliffeassociates.otter.common.data.model.ContentType
