+++ jvm/workbookapp/src/integration-test/kotlin/integrationtest/projects/TestRcImport.kt
@@ -1,4 +1,4 @@
-package integrationtest.rcimport
+package integrationtest.projects
 
 import org.junit.Test
 import org.wycliffeassociates.otter.common.data.model.ContentType.*
